issue 427: Added expiration_time and expirable_data to CellStoreTrailerV3

+++ src/cc/Common/TimeInline.h
@@ -82,7 +82,7 @@ parse_ts(const char *ts) {
     ns = strtol(last+1, &last, 10);
   }
 
-  return timegm(&tv) * G + sec * G + ns;
+  return (int64_t)(timegm(&tv) * G + sec * G + ns);
 }
 
 } // namespace Hypertable
